refactor(GuildCard): refactored styles and interfaces to one single file, updated imports
