--- src/Components/GuildCard/GuildCardContent.styled.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const DaoIcon = styled.img`
-  height: 4rem;
-  width: 4rem;
-`;
