+++ src/Components/GuildCard/types.tsx
@@ -9,3 +9,16 @@ export interface GuildCardProps extends CardProps {
   ensName: string;
   data: any;
 }
+
+export interface GuildCardContentProps {
+  isLoading?: boolean;
+  ensName: string;
+  data: any;
+}
+
+export interface GuildCardHeaderProps {
+  isLoading?: boolean;
+  t: any;
+  numberOfActiveProposals: any;
+  numberOfMembers: any;
+}
